+++ output/outieee.c
@@ -255,7 +255,7 @@ static void ieee_cleanup(int debuginfo)
  * callback for labels
  */
 static void ieee_deflabel(char *name, int32_t segment,
-                          int32_t offset, int is_global, char *special)
+                          int64_t offset, int is_global, char *special)
 {
     /*
      * We have three cases:
@@ -1348,7 +1348,7 @@ static void dbgls_linnum(const char *lnfname, int32_t lineno, int32_t segto)
 
 }
 static void dbgls_deflabel(char *name, int32_t segment,
-                           int32_t offset, int is_global, char *special)
+                           int64_t offset, int is_global, char *special)
 {
     struct ieeeSection *seg;
